separate skypilot flavors into different folders
