+++ src/zenml/integrations/skypilot/flavors/__init__.py
@@ -11,32 +11,3 @@
 #  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
 #  or implied. See the License for the specific language governing
 #  permissions and limitations under the License.
-"""Skypilot integration flavors."""
-
-from zenml.integrations.skypilot.flavors.skypilot_orchestrator_aws_vm_flavor import (
-    SkypilotAWSOrchestratorConfig,
-    SkypilotAWSOrchestratorFlavor,
-    SkypilotAWSOrchestratorSettings,
-)
-from zenml.integrations.skypilot.flavors.skypilot_orchestrator_gcp_vm_flavor import (
-    SkypilotGCPOrchestratorConfig,
-    SkypilotGCPOrchestratorFlavor,
-    SkypilotGCPOrchestratorSettings,
-)
-from zenml.integrations.skypilot.flavors.skypilot_orchestrator_azure_vm_flavor import (
-    SkypilotAzureOrchestratorConfig,
-    SkypilotAzureOrchestratorFlavor,
-    SkypilotAzureOrchestratorSettings,
-)
-
-__all__ = [
-    "SkypilotAWSOrchestratorConfig",
-    "SkypilotAWSOrchestratorFlavor",
-    "SkypilotAWSOrchestratorSettings",
-    "SkypilotGCPOrchestratorConfig",
-    "SkypilotGCPOrchestratorFlavor",
-    "SkypilotGCPOrchestratorSettings",
-    "SkypilotAzureOrchestratorConfig",
-    "SkypilotAzureOrchestratorFlavor",
-    "SkypilotAzureOrchestratorSettings",
-]
