+++ src/zenml/integrations/vm_aws/orchestrators/skypilot_aws_vm_orchestrator.py
@@ -11,21 +11,19 @@
 #  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
 #  or implied. See the License for the specific language governing
 #  permissions and limitations under the License.
-"""Implementation of the a Skypilot-based GCP VM orchestrator."""
+"""Implementation of the a Skypilot based AWS VM orchestrator."""
 
+import os
 from typing import TYPE_CHECKING, Optional, Type, cast
 
 import sky
 
-from zenml.integrations.gcp.google_credentials_mixin import (
-    GoogleCredentialsMixin,
-)
-from zenml.integrations.skypilot.flavors.skypilot_orchestrator_gcp_vm_flavor import (
-    SkypilotGCPOrchestratorConfig,
-    SkypilotGCPOrchestratorSettings,
-)
 from zenml.integrations.skypilot.orchestrators.skypilot_base_vm_orchestrator import (
     SkypilotBaseOrchestrator,
+)
+from zenml.integrations.vm_aws.flavors.skypilot_orchestrator_aws_vm_flavor import (
+    SkypilotAWSOrchestratorConfig,
+    SkypilotAWSOrchestratorSettings,
 )
 from zenml.logger import get_logger
 
@@ -34,16 +32,16 @@ if TYPE_CHECKING:
 
 logger = get_logger(__name__)
 
+ENV_AWS_PROFILE = "AWS_PROFILE"
 
-class SkypilotGCPOrchestrator(
-    SkypilotBaseOrchestrator, GoogleCredentialsMixin
-):
-    """Orchestrator responsible for running pipelines remotely in a VM on GCP.
+
+class SkypilotAWSOrchestrator(SkypilotBaseOrchestrator):
+    """Orchestrator responsible for running pipelines remotely in a VM on AWS.
 
     This orchestrator does not support running on a schedule.
     """
 
-    DEFAULT_INSTANCE_TYPE: str = "n1-standard-4"
+    DEFAULT_INSTANCE_TYPE: str = "t3.xlarge"
 
     @property
     def cloud(self) -> sky.clouds.Cloud:
@@ -52,16 +50,16 @@ class SkypilotGCPOrchestrator(
         Returns:
             A `sky.clouds.Cloud` instance.
         """
-        return sky.clouds.GCP()
+        return sky.clouds.AWS()
 
     @property
-    def config(self) -> SkypilotGCPOrchestratorConfig:
-        """Returns the `SkypilotGCPOrchestratorConfig` config.
+    def config(self) -> SkypilotAWSOrchestratorConfig:
+        """Returns the `SkypilotAWSOrchestratorConfig` config.
 
         Returns:
             The configuration.
         """
-        return cast(SkypilotGCPOrchestratorConfig, self._config)
+        return cast(SkypilotAWSOrchestratorConfig, self._config)
 
     @property
     def settings_class(self) -> Optional[Type["BaseSettings"]]:
@@ -70,12 +68,27 @@ class SkypilotGCPOrchestrator(
         Returns:
             The settings class.
         """
-        return SkypilotGCPOrchestratorSettings
+        return SkypilotAWSOrchestratorSettings
 
     def prepare_environment_variable(self, set: bool = True) -> None:
         """Set up Environment variables that are required for the orchestrator.
 
         Args:
             set: Whether to set the environment variables or not.
+
+        Raises:
+            ValueError: If no service connector is found.
         """
-        pass
+        connector = self.get_connector()
+        if connector is None:
+            raise ValueError(
+                "No service connector found. Please make sure to set up a connector "
+                "that is compatible with this orchestrator."
+            )
+        if set:
+            # The AWS connector creates a local configuration profile with the name computed from
+            # the first 8 digits of its UUID.
+            aws_profile = f"zenml-{str(connector.id)[:8]}"
+            os.environ[ENV_AWS_PROFILE] = aws_profile
+        else:
+            os.environ.pop(ENV_AWS_PROFILE, None)
