+++ src/zenml/integrations/vm_azure/flavors/__init__.py
@@ -11,3 +11,16 @@
 #  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
 #  or implied. See the License for the specific language governing
 #  permissions and limitations under the License.
+"""Skypilot integration flavors."""
+
+from zenml.integrations.vm_azure.flavors.skypilot_orchestrator_azure_vm_flavor import (
+    SkypilotAzureOrchestratorConfig,
+    SkypilotAzureOrchestratorFlavor,
+    SkypilotAzureOrchestratorSettings,
+)
+
+__all__ = [
+    "SkypilotAzureOrchestratorConfig",
+    "SkypilotAzureOrchestratorFlavor",
+    "SkypilotAzureOrchestratorSettings",
+]
