+++ src/zenml/integrations/vm_azure/orchestrators/__init__.py
@@ -1,4 +1,4 @@
-#  Copyright (c) ZenML GmbH 2023. All Rights Reserved.
+#  Copyright (c) ZenML GmbH 2024. All Rights Reserved.
 #
 #  Licensed under the Apache License, Version 2.0 (the "License");
 #  you may not use this file except in compliance with the License.
@@ -11,3 +11,16 @@
 #  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
 #  or implied. See the License for the specific language governing
 #  permissions and limitations under the License.
+"""Initialization of the Skypilot Azure ZenML orchestrator."""
+
+from zenml.integrations.skypilot.orchestrators.skypilot_base_vm_orchestrator import (  # noqa
+    SkypilotBaseOrchestrator,
+)
+from zenml.integrations.vm_azure.orchestrators.skypilot_azure_vm_orchestrator import (  # noqa
+    SkypilotAzureOrchestrator,
+)
+
+__all__ = [
+    "SkypilotBaseOrchestrator",
+    "SkypilotAzureOrchestrator",
+]
